+++ R/proprViz.R
@@ -104,6 +104,7 @@
 #'  a dynamic plot using the \code{plotly} package.
 #' @param col1,col2 A character vector. Specifies which nodes
 #'  to color \code{purple} or \code{orange}, respectively.
+#' @param d3 A boolean. Use \code{rgl} to plot 3D network.
 #'
 #' @importFrom stats var as.dist as.formula lm aov cutree prcomp dist kruskal.test
 #' @name visualize
@@ -585,7 +586,7 @@ snapshot <- function(rho, prompt = TRUE, plotly = FALSE){
 
 #' @rdname visualize
 #' @export
-cytescape <- function(object, col1, col2, prompt = TRUE){
+cytescape <- function(object, col1, col2, prompt = TRUE, d3 = FALSE){
 
   packageCheck("igraph")
 
@@ -609,15 +610,15 @@ cytescape <- function(object, col1, col2, prompt = TRUE){
 
   if(object@matrix[1, 1] == 0){
 
-    g <- migraph.color(g, sub$Partner[sub$rho < .1], sub$Pair[sub$rho < .1], "pink")
-    g <- migraph.color(g, sub$Partner[sub$rho < .05], sub$Pair[sub$rho < .05], "red")
+    g <- migraph.color(g, sub$Partner[sub$rho > 0], sub$Pair[sub$rho > 0], "greenyellow")
+    g <- migraph.color(g, sub$Partner[sub$rho < .05], sub$Pair[sub$rho < .05], "green")
 
   }else if(object@matrix[1, 1] == 1){
 
-    g <- migraph.color(g, sub$Partner[sub$rho > .95], sub$Pair[sub$rho > .95], "pink")
-    g <- migraph.color(g, sub$Partner[sub$rho > .98], sub$Pair[sub$rho > .98], "red")
-    g <- migraph.color(g, sub$Partner[sub$rho < -.95], sub$Pair[sub$rho < -.95], "lightblue")
-    g <- migraph.color(g, sub$Partner[sub$rho < -.98], sub$Pair[sub$rho < -.98], "blue")
+    g <- migraph.color(g, sub$Partner[sub$rho >= 0], sub$Pair[sub$rho >= 0], "greenyellow")
+    g <- migraph.color(g, sub$Partner[sub$rho > .98], sub$Pair[sub$rho > .98], "green")
+    g <- migraph.color(g, sub$Partner[sub$rho < 0], sub$Pair[sub$rho < 0], "tan3")
+    g <- migraph.color(g, sub$Partner[sub$rho < -.98], sub$Pair[sub$rho < -.98], "tan4")
     colnames(sub)[3] <- "phi"
 
   }else{
@@ -628,7 +629,17 @@ cytescape <- function(object, col1, col2, prompt = TRUE){
   if(!missing(col1)) g <- migraph.color(g, col1, col = "purple")
   if(!missing(col2)) g <- migraph.color(g, col2, col = "orange")
   g <- migraph.clean(g)
-  plot(g)
+
+  if(d3){
+
+    packageCheck("rgl")
+    coords <- igraph::layout_with_fr(g, dim = 3)
+    suppressWarnings(igraph::rglplot(g, layout = coords))
+
+  }else{
+
+    plot(g)
+  }
 
   return(sub)
 }